Place rocket data in rockets.js
